% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{MITIGRE}
\alias{MITIGRE}
\title{MITIGRE Application}
\usage{
MITIGRE(...)
}
\arguments{
\item{...}{No argument are needed to launch the application.}
}
\description{
An application to perform isotopic tracer modelling.
}
